backend/requirements.txt
asgiref==3.7.2
boto3==1.28.68
botocore==1.31.68
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.57.0
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-dotenv==1.4.2
django-import-export==3.3.1
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-storages==1.14.2
django-taggit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupPy==1.14
MarkupSafe==2.1.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
uritemplate==4.1.1
urllib3==2.0.7
xlrd==2.0.1
xlwt==1.3.0
